Make resolving its own service
